+++ plugins/inputs/socket_listener/socket_listener.go
@@ -114,7 +114,7 @@ func (ssl *streamSocketListener) read(c net.Conn) {
 		metrics, err := ssl.Parse(scnr.Bytes())
 		if err != nil {
 			ssl.AddError(fmt.Errorf("unable to parse incoming line: %s", err))
-			//TODO rate limit
+			// TODO rate limit
 			continue
 		}
 		for _, m := range metrics {
@@ -150,7 +150,7 @@ func (psl *packetSocketListener) listen() {
 		metrics, err := psl.Parse(buf[:n])
 		if err != nil {
 			psl.AddError(fmt.Errorf("unable to parse incoming packet: %s", err))
-			//TODO rate limit
+			// TODO rate limit
 			continue
 		}
 		for _, m := range metrics {
@@ -242,14 +242,17 @@ func (sl *SocketListener) Start(acc telegraf.Accumulator) error {
 		return fmt.Errorf("invalid service address: %s", sl.ServiceAddress)
 	}
 
-	if spl[0] == "unix" || spl[0] == "unixpacket" || spl[0] == "unixgram" {
+	protocol := spl[0]
+	addr := spl[1]
+
+	if protocol == "unix" || protocol == "unixpacket" || protocol == "unixgram" {
 		// no good way of testing for "file does not exist".
 		// Instead just ignore error and blow up when we try to listen, which will
 		// indicate "address already in use" if file existed and we couldn't remove.
-		os.Remove(spl[1])
+		os.Remove(addr)
 	}
 
-	switch spl[0] {
+	switch protocol {
 	case "tcp", "tcp4", "tcp6", "unix", "unixpacket":
 		var (
 			err error
@@ -258,17 +261,19 @@ func (sl *SocketListener) Start(acc telegraf.Accumulator) error {
 
 		tlsCfg, err := sl.ServerConfig.TLSConfig()
 		if err != nil {
-			return nil
+			return err
 		}
 
 		if tlsCfg == nil {
-			l, err = net.Listen(spl[0], spl[1])
+			l, err = net.Listen(protocol, addr)
 		} else {
-			l, err = tls.Listen(spl[0], spl[1], tlsCfg)
+			l, err = tls.Listen(protocol, addr, tlsCfg)
 		}
 		if err != nil {
 			return err
 		}
+
+		log.Printf("I! [inputs.socket_listener] Listening on %s://%s", protocol, l.Addr())
 
 		ssl := &streamSocketListener{
 			Listener:       l,
@@ -279,7 +284,7 @@ func (sl *SocketListener) Start(acc telegraf.Accumulator) error {
 		sl.Closer = ssl
 		go ssl.listen()
 	case "udp", "udp4", "udp6", "ip", "ip4", "ip6", "unixgram":
-		pc, err := net.ListenPacket(spl[0], spl[1])
+		pc, err := udpListen(protocol, addr)
 		if err != nil {
 			return err
 		}
@@ -288,9 +293,11 @@ func (sl *SocketListener) Start(acc telegraf.Accumulator) error {
 			if srb, ok := pc.(setReadBufferer); ok {
 				srb.SetReadBuffer(int(sl.ReadBufferSize.Size))
 			} else {
-				log.Printf("W! Unable to set read buffer on a %s socket", spl[0])
+				log.Printf("W! Unable to set read buffer on a %s socket", protocol)
 			}
 		}
+
+		log.Printf("I! [inputs.socket_listener] Listening on %s://%s", protocol, pc.LocalAddr())
 
 		psl := &packetSocketListener{
 			PacketConn:     pc,
@@ -300,14 +307,39 @@ func (sl *SocketListener) Start(acc telegraf.Accumulator) error {
 		sl.Closer = psl
 		go psl.listen()
 	default:
-		return fmt.Errorf("unknown protocol '%s' in '%s'", spl[0], sl.ServiceAddress)
+		return fmt.Errorf("unknown protocol '%s' in '%s'", protocol, sl.ServiceAddress)
 	}
 
-	if spl[0] == "unix" || spl[0] == "unixpacket" || spl[0] == "unixgram" {
+	if protocol == "unix" || protocol == "unixpacket" || protocol == "unixgram" {
 		sl.Closer = unixCloser{path: spl[1], closer: sl.Closer}
 	}
 
 	return nil
+}
+
+func udpListen(network string, address string) (net.PacketConn, error) {
+	switch network {
+	case "udp", "udp4", "udp6":
+		var addr *net.UDPAddr
+		var err error
+		var ifi *net.Interface
+		if spl := strings.SplitN(address, "%", 2); len(spl) == 2 {
+			address = spl[0]
+			ifi, err = net.InterfaceByName(spl[1])
+			if err != nil {
+				return nil, err
+			}
+		}
+		addr, err = net.ResolveUDPAddr(network, address)
+		if err != nil {
+			return nil, err
+		}
+		if addr.IP.IsMulticast() {
+			return net.ListenMulticastUDP(network, ifi, addr)
+		}
+		return net.ListenUDP(network, addr)
+	}
+	return net.ListenPacket(network, address)
 }
 
 func (sl *SocketListener) Stop() {