+++ plugins/inputs/vsphere/vsphere_test.go
@@ -341,8 +341,7 @@ func TestTimeout(t *testing.T) {
 	v.Timeout = internal.Duration{Duration: 1 * time.Nanosecond}
 	require.NoError(t, v.Start(nil)) // We're not using the Accumulator, so it can be nil.
 	defer v.Stop()
-	require.NoError(t, v.Gather(&acc))
-	require.True(t, len(acc.Errors) > 0, "Errors should not be empty here")
+	err = v.Gather(&acc)
 
 	// The accumulator must contain exactly one error and it must be a deadline exceeded.
 	require.Equal(t, 1, len(acc.Errors))