+++ plugins/parsers/registry.go
@@ -85,6 +85,9 @@ type Config struct {
 	// time format
 	JSONTimeFormat string `toml:"json_time_format"`
 
+	// default timezone
+	JSONTimezone string `toml:"json_timezone"`
+
 	// Authentication file for collectd
 	CollectdAuthFile string `toml:"collectd_auth_file"`
 	// One of none (default), sign, or encrypt
@@ -123,6 +126,7 @@ type Config struct {
 	GrokCustomPatterns     string   `toml:"grok_custom_patterns"`
 	GrokCustomPatternFiles []string `toml:"grok_custom_pattern_files"`
 	GrokTimezone           string   `toml:"grok_timezone"`
+	GrokUniqueTimestamp    string   `toml:"grok_unique_timestamp"`
 
 	//csv configuration
 	CSVColumnNames       []string `toml:"csv_column_names"`
@@ -152,6 +156,7 @@ func NewParser(config *Config) (Parser, error) {
 			config.JSONQuery,
 			config.JSONTimeKey,
 			config.JSONTimeFormat,
+			config.JSONTimezone,
 			config.DefaultTags)
 	case "value":
 		parser, err = NewValueParser(config.MetricName,
@@ -185,7 +190,8 @@ func NewParser(config *Config) (Parser, error) {
 			config.GrokNamedPatterns,
 			config.GrokCustomPatterns,
 			config.GrokCustomPatternFiles,
-			config.GrokTimezone)
+			config.GrokTimezone,
+			config.GrokUniqueTimestamp)
 	case "csv":
 		parser, err = newCSVParser(config.MetricName,
 			config.CSVHeaderRowCount,
@@ -275,6 +281,7 @@ func newJSONParser(
 	jsonQuery string,
 	timeKey string,
 	timeFormat string,
+	timezone string,
 	defaultTags map[string]string,
 ) Parser {
 	parser := &json.JSONParser{
@@ -285,17 +292,16 @@ func newJSONParser(
 		JSONQuery:      jsonQuery,
 		JSONTimeKey:    timeKey,
 		JSONTimeFormat: timeFormat,
+		JSONTimezone:   timezone,
 		DefaultTags:    defaultTags,
 	}
 	return parser
 }
 
-//Deprecated: Use NewParser to get a JSONParser object
 func newGrokParser(metricName string,
-	patterns []string,
-	nPatterns []string,
-	cPatterns string,
-	cPatternFiles []string, tZone string) (Parser, error) {
+	patterns []string, nPatterns []string,
+	cPatterns string, cPatternFiles []string,
+	tZone string, uniqueTimestamp string) (Parser, error) {
 	parser := grok.Parser{
 		Measurement:        metricName,
 		Patterns:           patterns,
@@ -303,6 +309,7 @@ func newGrokParser(metricName string,
 		CustomPatterns:     cPatterns,
 		CustomPatternFiles: cPatternFiles,
 		Timezone:           tZone,
+		UniqueTimestamp:    uniqueTimestamp,
 	}
 
 	err := parser.Compile()